Move routes to a separate file
